sistemas-monoliticos/src/modules/payments/repository/transaction.repository.ts
import { Transaction } from "../domain/transaction";
import { IPaymentGateway } from "../gateway/payment.gateway";
import { TransactionModel } from "../infra/transaction.model";

class TransactionRepository implements IPaymentGateway {
  async save(input: Transaction): Promise<Transaction> {
    await TransactionModel.create({
      id: input.id.id,
      orderId: input.orderId,
      amount: input.amount,
      status: input.status,
      createdAt: input.createdAt,
      updatedAt: input.updatedAt,
    });

    return new Transaction({
      id: input.id,
      orderId: input.orderId,
      amount: input.amount,
      status: input.status,
      createdAt: input.createdAt,
      updatedAt: input.updatedAt,
    });
  }
}

export { TransactionRepository }
